sp500-portfolio/requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
openpyxl>=3.1.2
requests>=2.31
scipy>=1.11
python-dotenv>=1.0
tqdm>=4.65
lxml>=4.9
html5lib>=1.1
